refactor(sdk): Move add_event_handler_impl to event_handler module
